refactor nodes to extract scene into its own module
